Extract the `HttpResponseHandler`.
